Safari-611.1.10/LayoutTests/fast/css/counters/content-021-expected.txt

This is the WebKit version of CSS Test: Content using a 'counters()' function with a string value.

PASS window.internals.counterValue(document.getElementById('div1')) is '0'
PASS window.internals.counterValue(document.getElementById('div2')) is '0.0'

TEST COMPLETE
